control-core/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
celery
bcrypt==4.0.1
redis
requests
alembic
python-dotenv
cryptography
jinja2
aiofiles
pytest
httpx
